a few more tests
